refactor: move code from command for better project consistency
