+++ jobMarket/services/analysis/posting_by_job_position/processing/__init__.py
@@ -1,0 +1,1 @@
+from .process import Pipeline
